Separated basicsSearch simplifications into their own files (#68)
