+++ test/simplifyExpression/basicsSearch/removeDivisionByOne.test.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const removeDivisionByOne = require('../../../lib/simplifyExpression/basicsSearch/removeDivisionByOne');
+const testSimplify = require('./testSimplify');
+
+describe('removeDivisionByOne', function() {
+  testSimplify('x/1', 'x', removeDivisionByOne);
+});
