+++ test/simplifyExpression/basicsSearch/removeExponentByOne.test.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const removeExponentByOne = require('../../../lib/simplifyExpression/basicsSearch/removeExponentByOne');
+const testSimplify = require('./testSimplify');
+
+describe('removeExponentByOne', function() {
+  testSimplify('x^1', 'x', removeExponentByOne);  
+});
